Refactor: through a generator function to generated style utils.
