--- src/themes/dark/index.ts
@@ -1,41 +1,0 @@
-import { generate } from '@ant-design/colors';
-import type { DerivativeFunc } from '@ant-design/cssinjs';
-
-import type { MapToken, PresetColorType, SeedToken } from '../../interface';
-import defaultAlgorithm from '../default';
-import { defaultPresetColors } from '../seed';
-import genColorMapToken from '../shared/genColorMapToken';
-import { generateColorPalettes, generateNeutralColorPalettes } from './colors';
-
-const derivative: DerivativeFunc<SeedToken, MapToken> = (token, mapToken) => {
-  const colorPalettes = Object.keys(defaultPresetColors)
-    .map((colorKey) => {
-      const colors = generate(token[colorKey as keyof PresetColorType], { theme: 'dark' });
-      return new Array(10).fill(1).reduce((prev, _, i) => {
-        prev[`${colorKey}-${i + 1}`] = colors[i];
-        prev[`${colorKey}${i + 1}`] = colors[i];
-        return prev;
-      }, {});
-    })
-    .reduce((prev, cur) => {
-      // biome-ignore lint/style/noParameterAssign: it is a reduce
-      prev = { ...prev, ...cur };
-      return prev;
-    }, {});
-
-  const mergedMapToken = mapToken ?? defaultAlgorithm(token);
-
-  return {
-    ...mergedMapToken,
-
-    // Dark tokens
-    ...colorPalettes,
-    // Colors
-    ...genColorMapToken(token, {
-      generateColorPalettes,
-      generateNeutralColorPalettes,
-    }),
-  };
-};
-
-export default derivative;
