--- src/themes/shared/genFontMapToken.ts
@@ -1,44 +1,0 @@
-import type { FontMapToken } from '../../interface';
-import genFontSizes from './genFontSizes';
-
-const genFontMapToken = (fontSize: number): FontMapToken => {
-  const fontSizePairs = genFontSizes(fontSize);
-  const fontSizes = fontSizePairs.map((pair) => pair.size);
-  const lineHeights = fontSizePairs.map((pair) => pair.lineHeight);
-
-  const fontSizeMD = fontSizes[1];
-  const fontSizeSM = fontSizes[0];
-  const fontSizeLG = fontSizes[2];
-  const lineHeight = lineHeights[1];
-  const lineHeightSM = lineHeights[0];
-  const lineHeightLG = lineHeights[2];
-
-  return {
-    fontSizeSM,
-    fontSize: fontSizeMD,
-    fontSizeLG,
-    fontSizeXL: fontSizes[3],
-
-    fontSizeHeading1: fontSizes[6],
-    fontSizeHeading2: fontSizes[5],
-    fontSizeHeading3: fontSizes[4],
-    fontSizeHeading4: fontSizes[3],
-    fontSizeHeading5: fontSizes[2],
-
-    lineHeight,
-    lineHeightLG,
-    lineHeightSM,
-
-    fontHeight: Math.round(lineHeight * fontSizeMD),
-    fontHeightLG: Math.round(lineHeightLG * fontSizeLG),
-    fontHeightSM: Math.round(lineHeightSM * fontSizeSM),
-
-    lineHeightHeading1: lineHeights[6],
-    lineHeightHeading2: lineHeights[5],
-    lineHeightHeading3: lineHeights[4],
-    lineHeightHeading4: lineHeights[3],
-    lineHeightHeading5: lineHeights[2],
-  };
-};
-
-export default genFontMapToken;
