+++ src/util/resetIcon.ts
@@ -1,0 +1,24 @@
+import { CSSObject } from '@ant-design/cssinjs';
+
+export const resetIcon = (): CSSObject => ({
+  display: 'inline-flex',
+  alignItems: 'center',
+  color: 'inherit',
+  fontStyle: 'normal',
+  lineHeight: 0,
+  textAlign: 'center',
+  textTransform: 'none',
+  // for SVG icon, see https://blog.prototypr.io/align-svg-icons-to-text-and-say-goodbye-to-font-icons-d44b3d7b26b4
+  verticalAlign: '-0.125em',
+  textRendering: 'optimizeLegibility',
+  '-webkit-font-smoothing': 'antialiased',
+  '-moz-osx-font-smoothing': 'grayscale',
+
+  '> *': {
+    lineHeight: 1,
+  },
+
+  svg: {
+    display: 'inline-block',
+  },
+});
