--- src/version/index.ts
@@ -1,1 +1,0 @@
-export default '1.0.0';
